+++ src/memory/static_data_area.h
@@ -8,7 +8,7 @@ extern int size;
 
 void push_global(int n);
 void pop_global(int n);
-void initialize_sda(size_t variable_count);
+void initialize_sda(uint32_t variable_count);
 void free_sda(void);
 void print_sda(void);
